split out multi view(s)
