+++ src/multiview-webpart/src/webparts/multiviews/components/multiview/IMultiviewsProps.ts
