+++ src/multiview-webpart/src/webparts/multiviews/components/multiview/welcome/IWelcomeProps.ts
